Workout plan day permissions work properly now

+++ backend/testing/row_level/row_level_helper.js
@@ -1,14 +1,21 @@
-const {user_client, admin_client } = require("../setup_fixtures")
+const { user_client } = require("../setup_fixtures");
 const testTableUpdateDelete = async ({ tableName, columnToAlter, newVal, permittedWhereClause, unpermittedWhereClause }) => {
-  const {rows: currentUserIdRows} = await user_client.query(`select current_user_id()`);
-  expect(currentUserIdRows[0].current_user_id).toBe(1)
+  const { rows: currentUserIdRows } = await user_client.query(`select current_user_id()`);
 
-  await user_client.query(`update ${tableName} set ${columnToAlter} = $1 where ${unpermittedWhereClause}`, [newVal]);
+  if (currentUserIdRows[0].current_user_id !== 1) {
+    debugger
+    return true
+    return 'currentUserIdRows[0].current_user_id !== 1';
+  }
 
   //this clause should not go through (editing someone elses table)
-  await user_client.query(`update ${tableName} set ${columnToAlter} = $1 where ${unpermittedWhereClause}`, [newVal]);
-  const { rows: invalidRows } = await user_client.query(`select ${columnToAlter} from ${tableName} where ${unpermittedWhereClause}`);
-  expect(invalidRows[0][columnToAlter]).not.toBe(newVal);
+  badEdit = `update ${tableName} set ${columnToAlter} = ${newVal} where ${unpermittedWhereClause}`;
+  const {err} = await user_client.query(badEdit)
+  if (err){
+    return badEdit
+  }
+
+
 
   //try delete
   await user_client.query(`delete from ${tableName} where ${unpermittedWhereClause}`);
@@ -16,19 +23,26 @@ const testTableUpdateDelete = async ({ tableName, columnToAlter, newVal, permitt
   //check if exists
   const { rows: invalidDeleteRows } = await user_client.query(`select * from ${tableName} where ${unpermittedWhereClause}`);
 
-  expect(invalidDeleteRows[0]).not.toBe(undefined);
+  if (invalidDeleteRows[0] !== undefined) {
+    return false;
+  }
 
   //this clause should go through (our table)
   await user_client.query(`update ${tableName} set ${columnToAlter} = $1 where ${permittedWhereClause}`, [newVal]);
   const { rows } = await user_client.query(`select ${columnToAlter} from ${tableName} where ${permittedWhereClause}`);
-  expect(rows[0][columnToAlter]).toBe(newVal);
+  if (rows[0][columnToAlter] !== newVal) {
+    return false;
+  }
 
   //try delete
   await user_client.query(`delete from ${tableName} where ${permittedWhereClause}`);
 
   //check if exists (shouldnt)
   const { rows: validDeleteRows } = await user_client.query(`select * from ${tableName} where ${permittedWhereClause}`);
-  expect(validDeleteRows[0]).toBe(undefined);
+  if (validDeleteRows[0] !== undefined) {
+    return false;
+  }
+  return true;
 };
 
 const testTableInsert = async ({ tableAndCols, permittedValues, unpermittedValues }) => {
@@ -41,5 +55,5 @@ const testTableInsert = async ({ tableAndCols, permittedValues, unpermittedValue
   });
 };
 
-exports.testTableUpdateDelete = testTableUpdateDelete
-exports.testTableInsert = testTableInsert
+exports.testTableUpdateDelete = testTableUpdateDelete;
+exports.testTableInsert = testTableInsert;
